requirements.txt
fastapi
python-multipart
uvicorn
openai
psycopg2
PyPDF2
numpy
python-docx
python-dotenv

